+++ clients/admin-ui/src/features/system/privacy-declarations/PrivacyDeclarationStep.tsx
@@ -30,6 +30,7 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
     allDataCategories: allEnabledDataCategories,
     allDataUses: allEnabledDataUses,
     allDataSubject: allEnabledDataSubjects,
+    cookies: system.cookies,
   };
 
   return (