+++ tests/ctl/core/test_api.py
@@ -2,6 +2,7 @@
 """Integration tests for the API module."""
 import json
 import typing
+from datetime import datetime
 from json import loads
 from typing import Dict, List
 
@@ -615,9 +616,12 @@ class TestSystemCreate:
 
         for field in SystemResponse.__fields__:
             system_val = getattr(system, field)
-            if isinstance(system_val, typing.Hashable):
+            if isinstance(system_val, typing.Hashable) and not isinstance(
+                system_val, datetime
+            ):
                 assert system_val == json_results[field]
         assert len(json_results["privacy_declarations"]) == 2
+        assert json_results["created_at"]
 
         for i, decl in enumerate(system.privacy_declarations):
             for field in PrivacyDeclarationResponse.__fields__:
@@ -1057,7 +1061,12 @@ class TestSystemUpdate:
             uses_profiling=True,
             legal_basis_for_profiling=["Authorised by law", "Contract"],
             does_international_transfers=True,
-            legal_basis_for_transfers=["Adequacy Decision", "BCRs"],
+            legal_basis_for_transfers=[
+                "Adequacy Decision",
+                "BCRs",
+                "Supplementary Measures",
+                "Unknown legal basis",
+            ],
             requires_data_protection_assessments=True,
             dpa_location="https://www.example.com/dpa",
             dpa_progress="pending",
@@ -1487,7 +1496,12 @@ class TestSystemUpdate:
         assert system.uses_profiling is True
         assert system.legal_basis_for_profiling == ["Authorised by law", "Contract"]
         assert system.does_international_transfers is True
-        assert system.legal_basis_for_transfers == ["Adequacy Decision", "BCRs"]
+        assert system.legal_basis_for_transfers == [
+            "Adequacy Decision",
+            "BCRs",
+            "Supplementary Measures",
+            "Unknown legal basis",
+        ]
         assert system.requires_data_protection_assessments is True
         assert system.dpa_location == "https://www.example.com/dpa"
         assert system.dpa_progress == "pending"
@@ -1525,10 +1539,13 @@ class TestSystemUpdate:
         json_results = result.json()
         for field in SystemResponse.__fields__:
             system_val = getattr(system, field)
-            if isinstance(system_val, typing.Hashable):
+            if isinstance(system_val, typing.Hashable) and not isinstance(
+                system_val, datetime
+            ):
                 assert system_val == json_results[field]
         assert len(json_results["privacy_declarations"]) == 1
         assert json_results["data_stewards"] == []
+        assert json_results["created_at"]
 
         for i, decl in enumerate(system.privacy_declarations):
             for field in PrivacyDeclarationResponse.__fields__: