+++ tests/ops/api/v1/endpoints/test_storage_endpoints.py
@@ -504,6 +504,7 @@ class TestGetStorageConfigs:
                 }
             ],
             "page": 1,
+            "pages": 1,
             "size": PAGE_SIZE,
             "total": 1,
         }
@@ -679,6 +680,7 @@ class TestGetDefaultStorageConfigs:
                 }
             ],
             "page": 1,
+            "pages": 1,
             "size": PAGE_SIZE,
             "total": 1,
         }